........
  r33845 | res2002 | 2010-03-29 21:18:42 -0400 (Mon, 29 Mar 2010) | 1 line
  
  Use uintptr_t for ThreadID (thread ID is apparently a pointer type on Snow Leopard)
........


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/release/V1.4@34876 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/threading/thread.h
@@ -67,7 +67,7 @@ namespace Threading
   };
 
   /// OS specific thread identifier
-  typedef unsigned int ThreadID;
+  typedef uintptr_t ThreadID;
 
 }
 }
